Add allow_headers and hostname to config
Each one of the allowed headers will be passed to drakov as `--header header`.
Hardcoded headers are removed from the drakov service.

Default port is set to 8082 in `Config` to remove the hardcoded configuration
in `lib/blueprint_agreement.rb`.

Normalize all config options in `Config` module to use `@@` instead of `@`.